Refactor cyberway contracts to simplify import changes from EOS. #479
