+++ libraries/chain/CMakeLists.txt
@@ -4,6 +4,7 @@ configure_file(${CMAKE_CURRENT_SOURCE_DIR}/genesis_state_root_key.cpp.in ${CMAKE
 file(GLOB HEADERS "include/eosio/chain/*.hpp"
                   "include/eosio/chain/webassembly/*.hpp"
                   "include/cyberway/chaindb/*.hpp"
+                  "include/cyberway/chain/*.hpp"
                   "include/cyberway/genesis/*.hpp"
                   "${CMAKE_CURRENT_BINARY_DIR}/include/eosio/chain/core_symbol.hpp" )
 
@@ -53,7 +54,8 @@ add_library( eosio_chain
              chaindb/abi_info.cpp
              chaindb/ram_calculator.cpp
 
-             domain_name.cpp
+             cyberway/domain_name.cpp
+             cyberway/cyberway_contract.cpp
              genesis/genesis_import.cpp
              ${CHAINDB_SRCS}
