+++ libraries/chain/include/cyberway/chain/domain_name.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <string>
 
-namespace eosio { namespace chain {
+namespace cyberway { namespace chain {
 
 using domain_name = std::string;    // TODO: separate class to embed restrictions into deserializer
 using username = std::string;
@@ -9,5 +9,5 @@ using username = std::string;
 void validate_domain_name(const domain_name& n);
 void validate_username(const username& n);
 
-} } // eosio::chain
+} } // cyberway::chain
