+++ libraries/chain/include/eosio/chain/config.hpp
@@ -49,11 +49,6 @@ const static uint64_t owner_name  = N(owner);
 const static uint64_t eosio_any_name = N(cyber.any);
 const static uint64_t eosio_code_name = N(cyber.code);
 
-// TODO: requestbw
-//const static uint64_t request_bw_action = N(requestbw);
-//const static uint64_t approve_bw_action = N(approvebw);
-const static uint64_t provide_bw_action = N(providebw);
-
 const static int      block_interval_ms = 3000;
 const static int      block_interval_us = block_interval_ms*1000;
 const static uint64_t block_timestamp_epoch = 946684800000ll; // epoch is year 2000.
