+++ libraries/chain/include/eosio/chain/contract_types.hpp
@@ -138,101 +138,6 @@ struct canceldelay {
    }
 };
 
-struct providebw final {
-    account_name    provider;
-    account_name    account;
-
-    providebw() = default;
-    providebw(const account_name& provider, const account_name& account)
-    : provider(provider), account(account)
-    {}
-
-    static account_name get_account() {
-        return config::system_account_name;
-    }
-
-    static action_name get_name() {
-        return config::provide_bw_action;
-    }
-};
-
-// TODO: requestbw
-//struct requestbw {
-//    account_name    provider;
-//    account_name    account;
-//
-//    requestbw() = default;
-//    requestbw(const account_name& provider, const account_name& account)
-//    : provider(provider), account(account)
-//    {}
-//
-//    static account_name get_account() {
-//        return config::system_account_name;
-//    }
-//
-//    static action_name get_name() {
-//        return config::request_bw_action;
-//    }
-//};
-
-struct provideram final {
-    account_name provider;
-    account_name account;
-
-    provideram() = default;
-    provideram(const account_name& provider, const account_name& account)
-    : provider(provider), account(account) {
-    }
-
-    static account_name get_account() {
-        return config::system_account_name;
-    }
-
-    static action_name get_name() {
-        return N(provideram);
-    }
-};
-
-
-// it's ugly, but removes boilerplate. TODO: write better
-#define SYS_ACTION_STRUCT(NAME) struct NAME { \
-    NAME() = default; \
-    static account_name get_account() { return config::domain_account_name; } \
-    static action_name get_name()     { return N(NAME); }
-#define SYS_ACTION_STRUCT_END };
-
-SYS_ACTION_STRUCT(newdomain)
-   account_name creator;
-   domain_name name;
-SYS_ACTION_STRUCT_END;
-
-SYS_ACTION_STRUCT(passdomain)
-   account_name from;
-   account_name to;
-   domain_name name;
-SYS_ACTION_STRUCT_END;
-
-SYS_ACTION_STRUCT(linkdomain)
-   account_name owner;
-   account_name to;
-   domain_name name;
-SYS_ACTION_STRUCT_END;
-
-SYS_ACTION_STRUCT(unlinkdomain)
-   account_name owner;
-   domain_name name;
-SYS_ACTION_STRUCT_END;
-
-SYS_ACTION_STRUCT(newusername)
-   account_name creator;
-   account_name owner;
-   username name;
-SYS_ACTION_STRUCT_END;
-
-#undef SYS_ACTION_STRUCT
-#undef SYS_ACTION_STRUCT_END
-
-
 struct onerror {
    uint128_t      sender_id;
    bytes          sent_trx;
@@ -249,24 +154,6 @@ struct onerror {
    }
 };
 
-// TODO: requestbw
-//struct approvebw {
-//    account_name    account;
-//
-//    approvebw() = default;
-//    approvebw(const account_name& account)
-//    : account(account)
-//    {}
-//
-//    static account_name get_account() {
-//       return config::system_account_name;
-//    }
-//
-//    static action_name get_name() {
-//        return config::approve_bw_action;
-//    }
-//};
-
 } } /// namespace eosio::chain
 
 FC_REFLECT( eosio::chain::newaccount                       , (creator)(name)(owner)(active) )
@@ -276,16 +163,5 @@ FC_REFLECT( eosio::chain::updateauth                       , (account)(permissio
 FC_REFLECT( eosio::chain::deleteauth                       , (account)(permission) )
 FC_REFLECT( eosio::chain::linkauth                         , (account)(code)(type)(requirement) )
 FC_REFLECT( eosio::chain::unlinkauth                       , (account)(code)(type) )
-FC_REFLECT( eosio::chain::providebw                        , (provider)(account) )
-//TODO: requestbw
-//FC_REFLECT( eosio::chain::requestbw                        , (provider)(account) )
-//FC_REFLECT( eosio::chain::approvebw                        , (account) )
-FC_REFLECT( eosio::chain::provideram                       , (provider)(account) )
 FC_REFLECT( eosio::chain::canceldelay                      , (canceling_auth)(trx_id) )
 FC_REFLECT( eosio::chain::onerror                          , (sender_id)(sent_trx) )
-
-FC_REFLECT(eosio::chain::newdomain,    (creator)(name))
-FC_REFLECT(eosio::chain::passdomain,   (from)(to)(name))
-FC_REFLECT(eosio::chain::linkdomain,   (owner)(to)(name))
-FC_REFLECT(eosio::chain::unlinkdomain, (owner)(name))
-FC_REFLECT(eosio::chain::newusername,  (creator)(owner)(name))
